add cookiecutter config
Add a basic cookiecutter configuration which specifies what templatable
variables there are in the project.

project_{name,slug} are used to populate the human-readable name of the
project and a file-name friendly version of the project name.

project_module is the name of the Django project module. I've never been
very happy with having to come up with two similar but different names
to refer to the django project and (usually) the single application
within it so I'm embracing the idea of having "project" be the default.

application_{name,slug} will be used to create the primary application.