settings.base: fix BASE_DIR value
The ``BASE_DIR`` value was left at that generated by the default Django
project template but that assumed that the settings file was located
within the project root. In our template we locate settings files within
their own ``settings`` module. This means that ``BASE_DIR`` points to
``PROJECT_ROOT`` and not the directory which contains ``PROJECT_ROOT``
as it should.

Add one more ``os.path.dirname()`` wrapper to the calculation of
``BASE_DIR`` to fix this.

This issue was first found in uisautomation/sms-webapp.